APM-2060 Refactor get_env into conftest
